+++ src/Passthru/ContextInterface.php
@@ -13,6 +13,8 @@ interface ContextInterface
 {
     public function getResponse(): ResponseInterface;
 
+    public function getStatusCode(): int;
+
     public function getResponseBody(): ?string;
 
     public function getInfo(?string $key = null): mixed;